+++ SRC/Public/Abstractions/IInjector.cs
@@ -8,8 +8,6 @@ using System.Diagnostics.CodeAnalysis;
 
 namespace Solti.Utils.DI
 {
-    using Internals;
-
     /// <summary>
     /// Provides the mechanism for injecting resources.
     /// </summary>
@@ -24,13 +22,5 @@ namespace Solti.Utils.DI
         /// <exception cref="ServiceNotFoundException">The service or one or more dependencies could not be found.</exception>
         [SuppressMessage("Naming", "CA1716:Identifiers should not match keywords", Justification = "The identifier won't confuse the users of the API.")]
         object Get(Type iface, string name = null);
-
-        /// <summary>
-        /// Gets the <see cref="Lifetime"/> of the given service (type).
-        /// </summary>
-        /// <param name="iface">>The "id" of the service.</param>
-        /// <param name="name">The (optional) name of the service.</param>
-        /// <returns>The <see cref="Lifetime"/> of the service if it is producible, null otherwise.</returns>
-        Lifetime? LifetimeOf(Type iface, string name = null);
     }
 }